fix: Replace before filters with get handlers
The before filters were very broad ( matching all get requests), as a result non spa-web-app routes needed to be excluded, resulting in losts of branches.
Sprak java uses a route handler design that proceses the requests in the order in which the handlers are defined.
By rearranging the order of the handlers to match spesific routes before more general routes ( like most frameworks do ) the overly broad before matchers can be removed.

Closes #1081