fix: fetch missing tree parts work-around
Due to the lack of recursive query posabilities the partial tree can not be reconstructed from the selected ontology items.
This change works around this missing feature by fetching up to 8 levels up (from the selected ontology item).
This is sufficient for the current ontologies used but does not slove te general case.