fix: default of table explorer to be 'canView' (#2770)
* fix: default of table explorer to be 'canView'

We have issue that we now can disable 'canView' to accomodate when user has 'aggregate' permission only. But default was 'false' which lead to random errors in other places using the table explorer. With this fix we disable that, however we must make sure this doesn't lead to errors in case user has only aggregate.

fix: add flip to wrapping component

---------

Co-authored-by: connoratrug <c.stroomberg@rug.nl>
Co-authored-by: connoratrug <47183404+connoratrug@users.noreply.github.com>