fix: directory fixes
The following issues have been fixed:

The contact and head information is displayed on the biobank (side) card.
The AlsoKnownIn urls are displayed on biobank, collection and network cards.
The Quality Standards information is being retrieved from QualityStandards table and is displayed on the according cards.
Age unit and range is displayed on the collection card.