feat(analytics, ui): event based analytics service plus (pre alpha)tailwind UI for the analytics config (#4031)
Analytics

Allow admin user to setup analytics events. When the enduser of the web interface executes an action ( for example clicks on the 'contact' button) a signal is send to a analytics service ( third party; site-improve, google, .. or emx2 analytics endpoint). The app maintainer can configure which events are send in response to what ui element the user interacts with ( initial 'click' event only). The maintainer can configure these events on a per schema basis.



Tailwind (analytics) ui (pre alpha)
- ui path: /apps/ui/#/
- login as admin to update analytics settings
- analytics path: /ui/[schema-name]/analytics