fix: add workaround for java-spark route matching limitations
On the backend the java spark staticFiles matcher run before all other matches ( this is non configurable, unless you turn of staticFiles)
If a resource is found ( nuxt spa mode creates a path stuct the index.html files for the (non)dynamic routes ) java-spark only checks for / router ( home-page of route ending in '/' , only in these cases it tries to return a index.html otherwise it tries to guess the resp.mimeType from the file name, in case of a non homepage route ( i.e. apps/ui/login ) java-spark asssumes the filename to be 'login' ( instead of the correct index.html) and then defaults the return mineType to the incorrect 'application/octet-stream' resulting in a download action from the browser. this change is a work around for the java-spark limitation, by appending the '/' the index.html is assumed by spark