fix: 5717 clear selection on update
Use null instead of undefined on clearing a selection on update.
When the form modal is transformed into a formData object to be send with the mutation, the fields that need to be cleared need to be included in the formData. Setting the values to null instead of undefined ensures the fields are included in the formData.

Closes #5717